move degreeworks scripts
